Basic structure of kalloc/physical memmory manager done
