+++ kernel/arch/x86_64/make.config
@@ -19,4 +19,5 @@ KERNEL_ARCH_OBJS= \
     $(ARCHDIR)/cpu/cpudet.o \
     $(ARCHDIR)/cpu/apic.o \
     $(ARCHDIR)/mem/hhdm.o \
-    $(ARCHDIR)/mem/mem.o
+    $(ARCHDIR)/mem/mem.o \
+    $(ARCHDIR)/mem/pmm.o
